Restructure JointType code into several files. Consistently use propagate_inbounds and inline.
