+++ src/RigidBodyDynamics.jl
@@ -7,7 +7,9 @@ using Rotations
 using TypeSortedCollections
 using DocStringExtensions
 using Reexport
-import Base.Iterators: filter, flatten
+
+using Base.Iterators: filter, flatten
+using Base: @propagate_inbounds
 
 # mechanism-related types
 export
@@ -158,8 +160,8 @@ export
     segments
 
 include("custom_collections.jl")
-include("graphs/Graphs.jl")
-include("spatial/Spatial.jl")
+include(joinpath("graphs", "Graphs.jl"))
+include(joinpath("spatial", "Spatial.jl"))
 include("contact.jl")
 include("pdcontrol.jl")
 
@@ -173,7 +175,7 @@ import .Spatial: rotation, translation, transform, center_of_mass, newton_euler,
 
 include("util.jl")
 include("joint.jl")
-include("joint_types.jl")
+include(joinpath("joint_types", "joint_types.jl"))
 include("rigid_body.jl")
 include("mechanism.jl")
 include("mechanism_modification.jl")
@@ -184,7 +186,7 @@ include("mechanism_algorithms.jl")
 include("ode_integrators.jl")
 include("simulate.jl")
 
-include("urdf/URDF.jl")
+include(joinpath("urdf", "URDF.jl"))
 @reexport using .URDF
 
 # import these for MechanismGeometries compatibility. TODO: stop importing these after updating MechanismGeometries.
