+++ src/mechanism_state.jl
@@ -1,18 +1,18 @@
 ## BodyDict, JointDict
 const BodyDict{V} = IndexDict{BodyID, Base.OneTo{BodyID}, V}
 const BodyCacheDict{V} = CacheIndexDict{BodyID, Base.OneTo{BodyID}, V}
-Base.@propagate_inbounds Base.getindex(d::AbstractIndexDict{BodyID}, key::RigidBody) = d[BodyID(key)]
-Base.@propagate_inbounds Base.setindex!(d::AbstractIndexDict{BodyID}, value, key::RigidBody) = d[BodyID(key)] = value
+@propagate_inbounds Base.getindex(d::AbstractIndexDict{BodyID}, key::RigidBody) = d[BodyID(key)]
+@propagate_inbounds Base.setindex!(d::AbstractIndexDict{BodyID}, value, key::RigidBody) = d[BodyID(key)] = value
 
 const JointDict{V} = IndexDict{JointID, Base.OneTo{JointID}, V}
 const JointCacheDict{V} = CacheIndexDict{JointID, Base.OneTo{JointID}, V}
-Base.@propagate_inbounds Base.getindex(d::AbstractIndexDict{JointID}, key::Joint) = d[JointID(key)]
-Base.@propagate_inbounds Base.setindex!(d::AbstractIndexDict{JointID}, value, key::Joint) = d[JointID(key)] = value
+@propagate_inbounds Base.getindex(d::AbstractIndexDict{JointID}, key::Joint) = d[JointID(key)]
+@propagate_inbounds Base.setindex!(d::AbstractIndexDict{JointID}, value, key::Joint) = d[JointID(key)] = value
 
 
 ## SegmentedVector method overloads
-Base.@propagate_inbounds Base.getindex(v::SegmentedVector{JointID}, id::JointID) = v.segments[id]
-Base.@propagate_inbounds Base.getindex(v::SegmentedVector{JointID}, joint::Joint) = v[JointID(joint)]
+@propagate_inbounds Base.getindex(v::SegmentedVector{JointID}, id::JointID) = v.segments[id]
+@propagate_inbounds Base.getindex(v::SegmentedVector{JointID}, joint::Joint) = v[JointID(joint)]
 function SegmentedVector(parent::AbstractVector{T}, joints::AbstractVector{<:Joint}, viewlengthfun) where T
     SegmentedVector{JointID, T, Base.OneTo{JointID}}(parent, joints, viewlengthfun)
 end
@@ -189,8 +189,8 @@ Base.broadcastable(x::MechanismState) = Ref(x)
 Base.show(io::IO, ::MechanismState{X, M, C}) where {X, M, C} = print(io, "MechanismState{$X, $M, $C, …}(…)")
 
 modcount(state::MechanismState) = state.modcount
-Base.@propagate_inbounds predsucc(id::JointID, state::MechanismState) = state.predecessor_and_successor_ids[id]
-Base.@propagate_inbounds successorid(id::JointID, state::MechanismState) = last(state.predecessor_and_successor_ids[id])
+@propagate_inbounds predsucc(id::JointID, state::MechanismState) = state.predecessor_and_successor_ids[id]
+@propagate_inbounds successorid(id::JointID, state::MechanismState) = last(state.predecessor_and_successor_ids[id])
 
 """
 $(SIGNATURES)
@@ -560,7 +560,7 @@ $(SIGNATURES)
 Return the joint transform for the given joint, i.e. the transform from
 `frame_after(joint)` to `frame_before(joint)`.
 """
-Base.@propagate_inbounds function joint_transform(state::MechanismState, joint::Union{<:Joint, JointID}, safe=true)
+@propagate_inbounds function joint_transform(state::MechanismState, joint::Union{<:Joint, JointID}, safe=true)
     safe && update_transforms!(state)
     state.joint_transforms[joint]
 end
@@ -571,7 +571,7 @@ Return the joint twist for the given joint, i.e. the twist of
 `frame_after(joint)` with respect to `frame_before(joint)`, expressed in the
 root frame of the mechanism.
 """
-Base.@propagate_inbounds function twist(state::MechanismState, joint::Union{<:Joint, JointID}, safe=true)
+@propagate_inbounds function twist(state::MechanismState, joint::Union{<:Joint, JointID}, safe=true)
     safe && update_joint_twists!(state)
     state.joint_twists[joint]
 end
@@ -583,7 +583,7 @@ Return the bias acceleration across the given joint, i.e. the spatial accelerati
 of `frame_after(joint)` with respect to `frame_before(joint)`, expressed in the
 root frame of the mechanism when all joint accelerations are zero.
 """
-Base.@propagate_inbounds function bias_acceleration(state::MechanismState, joint::Union{<:Joint, JointID}, safe=true)
+@propagate_inbounds function bias_acceleration(state::MechanismState, joint::Union{<:Joint, JointID}, safe=true)
     safe && update_joint_bias_accelerations!(state)
     state.joint_bias_accelerations[joint]
 end
@@ -594,7 +594,7 @@ $(SIGNATURES)
 Return the transform from `default_frame(body)` to the root frame of the
 mechanism.
 """
-Base.@propagate_inbounds function transform_to_root(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function transform_to_root(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     safe && update_transforms!(state)
     state.transforms_to_root[body]
 end
@@ -605,7 +605,7 @@ $(SIGNATURES)
 Return the twist of `default_frame(body)` with respect to the root frame of the
 mechanism, expressed in the root frame.
 """
-Base.@propagate_inbounds function twist_wrt_world(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function twist_wrt_world(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     safe && update_twists_wrt_world!(state)
     state.twists_wrt_world[body]
 end
@@ -618,7 +618,7 @@ i.e. the spatial acceleration of `default_frame(body)` with respect to the
 root frame of the mechanism, expressed in the root frame, when all joint
 accelerations are zero.
 """
-Base.@propagate_inbounds function bias_acceleration(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function bias_acceleration(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     safe && update_bias_accelerations_wrt_world!(state)
     state.bias_accelerations_wrt_world[body]
 end
@@ -629,7 +629,7 @@ $(SIGNATURES)
 Return the spatial inertia of `body` expressed in the root frame of the
 mechanism.
 """
-Base.@propagate_inbounds function spatial_inertia(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function spatial_inertia(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     safe && update_spatial_inertias!(state)
     state.inertias[body]
 end
@@ -640,7 +640,7 @@ $(SIGNATURES)
 Return the composite rigid body inertia `body` expressed in the root frame of the
 mechanism.
 """
-Base.@propagate_inbounds function crb_inertia(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function crb_inertia(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     safe && update_crb_inertias!(state)
     state.crb_inertias[body]
 end
@@ -833,21 +833,21 @@ end
 
 contact_states(state::MechanismState, body::Union{<:RigidBody, BodyID}) = state.contact_states[body]
 
-Base.@propagate_inbounds function newton_euler(state::MechanismState, body::Union{<:RigidBody, BodyID}, accel::SpatialAcceleration, safe=true)
+@propagate_inbounds function newton_euler(state::MechanismState, body::Union{<:RigidBody, BodyID}, accel::SpatialAcceleration, safe=true)
     inertia = spatial_inertia(state, body, safe)
     twist = twist_wrt_world(state, body, safe)
     newton_euler(inertia, accel, twist)
 end
 
-Base.@propagate_inbounds function momentum(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function momentum(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     spatial_inertia(state, body, safe) * twist_wrt_world(state, body, safe)
 end
 
-Base.@propagate_inbounds function momentum_rate_bias(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function momentum_rate_bias(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     newton_euler(state, body, bias_acceleration(state, body, safe), safe)
 end
 
-Base.@propagate_inbounds function kinetic_energy(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function kinetic_energy(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     kinetic_energy(spatial_inertia(state, body, safe), twist_wrt_world(state, body, safe))
 end
 
@@ -858,7 +858,7 @@ Return the gravitational potential energy in the given state, computed as the
 negation of the dot product of the gravitational force and the center
 of mass expressed in the `Mechanism`'s root frame.
 """
-Base.@propagate_inbounds function gravitational_potential_energy(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
+@propagate_inbounds function gravitational_potential_energy(state::MechanismState, body::Union{<:RigidBody, BodyID}, safe=true)
     inertia = spatial_inertia(body)
     m = inertia.mass
     m > 0 || return zero(cache_eltype(state))
@@ -928,7 +928,7 @@ function transform_to_root(state::MechanismState, frame::CartesianFrame3D, safe=
     tf
 end
 
-Base.@propagate_inbounds function statesum(fun::F, state::MechanismState, itr, safe=true; init) where F
+@propagate_inbounds function statesum(fun::F, state::MechanismState, itr, safe=true; init) where F
     ret = init
     for x in itr
         ret += fun(state, x, safe)
@@ -936,28 +936,28 @@ Base.@propagate_inbounds function statesum(fun::F, state::MechanismState, itr, s
     ret
 end
 
-Base.@propagate_inbounds function momentum(state::MechanismState, body_itr)
+@propagate_inbounds function momentum(state::MechanismState, body_itr)
     T = cache_eltype(state)
     update_twists_wrt_world!(state)
     update_spatial_inertias!(state)
     statesum(momentum, state, body_itr, false, init=zero(Momentum{T}, root_frame(state.mechanism)))
 end
 
-Base.@propagate_inbounds function momentum_rate_bias(state::MechanismState, body_itr)
+@propagate_inbounds function momentum_rate_bias(state::MechanismState, body_itr)
     T = cache_eltype(state)
     update_bias_accelerations_wrt_world!(state)
     update_spatial_inertias!(state)
     statesum(momentum_rate_bias, state, body_itr, false, init=zero(Wrench{T}, root_frame(state.mechanism)))
 end
 
-Base.@propagate_inbounds function kinetic_energy(state::MechanismState, body_itr)
+@propagate_inbounds function kinetic_energy(state::MechanismState, body_itr)
     T = cache_eltype(state)
     update_twists_wrt_world!(state)
     update_spatial_inertias!(state)
     statesum(kinetic_energy, state, body_itr, false, init=zero(T))
 end
 
-Base.@propagate_inbounds function gravitational_potential_energy(state::MechanismState, body_itr)
+@propagate_inbounds function gravitational_potential_energy(state::MechanismState, body_itr)
     T = cache_eltype(state)
     update_transforms!(state)
     statesum(gravitational_potential_energy, state, body_itr, false, init=zero(T))
